requirements.txt
# Python 3.8+
pandas
python-dotenv
sqlalchemy
pyodbc
psutil